\name{pyrithione}
\docType{data}
\alias{pyrithione}
\title{Cytotoxicity data for different pyrithionates and related species}
\description{
 This data shows the cytotoxicity of pyrithione salts as well as the free
 pyrithione, its oxidation product and some other related compounds to the
 IPC-81 cell line.
}
\usage{data(pyrithione)}
\format{
  A dataframe containing the data as required for the \code{\link{drfit}} function.
}
\source{
  Doose C, Ranke J, Stock F, Bottin-Weber U, Jastorff B (2004)
  Structure-activity relationships of pyrithiones - IPC-81 toxicity tests with
  antifouling biocide zinc pyrithione and structural analogues. Green Chemistry
  6(5) 259-266 }
\keyword{datasets}
