\name{XY}
\docType{data}
\alias{XY}
\title{Dose-Response data for two substances X and Y}
\description{
  This is just a sample Lemna growth rate data set for two substances 
  arbitrarily named X and Y.
}
\usage{data(antifoul)}
\format{
  A dataframe containing dose (concentration) and response data, as well as
  control values where the dose is zero.
}
\source{
  \url{http://www.uft.uni-bremen.de/chemie}
}
\keyword{datasets}
